+++ my-code/src/components/detail/detail.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import {Link, useParams} from "react-router-dom";
 
 import { MovieData } from "../../interfaces/detail";
-import fetchApiData from "../../utils/fetchData";
-import { resultIsError } from "../../utils/checkResultAsError";
+import fetchApiData from "../../utils";
+import { resultIsError } from "../../utils";
 
 import Spinner from "../spinner/spinner";
 import DetailColumns from "../detailColumns";
@@ -15,6 +15,7 @@ import imdb from '../../assets/2.Logos/logo-imdb.svg';
 
 import './detail.css';
 import ErrorMessage from "../errorMessage";
+import Label from "../label";
 
 export default function Detail() {
     const [loading, setLoading] = useState(false);
@@ -32,10 +33,10 @@ export default function Detail() {
         const url = `http://www.omdbapi.com/?apikey=b02d2b50&i=${id}`;
         if (mounted) {
             (async () => {
-                const { data, error } = await fetchApiData<MovieData>(url);
+                const { data, error: fetchError } = await fetchApiData<MovieData>(url);
                 setLoading(false);
-                if (error) {
-                    setError(error);
+                if (fetchError) {
+                    setError(fetchError);
                 } else if (resultIsError(data)) {
                     setError(data.Error);
                 } else {
@@ -49,60 +50,62 @@ export default function Detail() {
     }, [paramsObj]);
 
     const backFillToggle = () => {
-        setBackFill(!backFill);
+        setBackFill(value => !value);
     }
 
     return (
         <div className="detail-wrapper">
             {loading ?
                 <Spinner />
-            : error ?
+            : null}
+            {error ?
                 <ErrorMessage message={`There was an error loading the movie: ${error}`} />
-            :
-            <>
-                <div className="icon-back">
-                    <Link to='/'>
-                        <span onMouseEnter={backFillToggle} onMouseLeave={backFillToggle}>
-                            <Back fill={backFill ? 'white' : '#7A8C99'} />
-                        </span>
-                    </Link>
-                </div>
-                <div className="detail-grid">
-                    <div className="text">
-                        <div className="top">
-                            <span id="runtime" className="runtime">{movie?.Runtime} &middot; {movie?.Year} &middot;</span>
-                            <span className="rated">{movie?.Rated}</span>
+                : null}
+            {!loading && !error ?
+                <>
+                    <div className="icon-back">
+                        <Link to='/'>
+                            <span onMouseEnter={backFillToggle} onMouseLeave={backFillToggle}>
+                                <Back fill={backFill ? 'white' : '#7A8C99'} />
+                            </span>
+                        </Link>
+                    </div>
+                    <div className="detail-grid">
+                        <div className="text">
+                            <div className="top">
+                                <span id="runtime" className="runtime">{movie?.Runtime} &middot; {movie?.Year} &middot;</span>
+                                <span className="rated">{movie?.Rated}</span>
+                            </div>
+                            <div className="title">
+                                <h1>
+                                    {movie?.Title}
+                                </h1>
+                            </div>
+                            <div className="btn-section">
+                                <Label className="imdb" image={imdb} text={movie?.Ratings[0]?.Value || 'N/A'} />
+                                <Label className="rotten" image={rotten} text={movie?.Ratings[1]?.Value || 'N/A'} />
+                                <Button />
+                            </div>
+                            <div className="plot">
+                                <h4 className="title">
+                                    Plot
+                                </h4>
+                                <h4 className="content">
+                                    {movie?.Plot}
+                                </h4>
+                            </div>
+                            <div className="bottom">
+                                <DetailColumns title='Cast' data={movie?.Actors} />
+                                <DetailColumns title='Genre' data={movie?.Genre} />
+                                <DetailColumns title='Director' data={movie?.Director} />
+                            </div>
                         </div>
-                        <div className="title">
-                            <h1>
-                                {movie?.Title}
-                            </h1>
-                        </div>
-                        <div className="btn-section">
-                            <Button className="imdb" image={imdb} text={movie?.Ratings[0]?.Value || 'N/A'}/>
-                            <Button className="rotten" image={rotten} text={movie?.Ratings[1]?.Value || 'N/A'} />
-                            <Button className="favourite" favouriteBtn/>
-                        </div>
-                        <div className="plot">
-                            <h4 className="title">
-                                Plot
-                            </h4>
-                            <h4 className="content">
-                                {movie?.Plot}
-                            </h4>
-                        </div>
-                        <div className="bottom">
-                            <DetailColumns title='Cast' data={movie?.Actors} />
-                            <DetailColumns title='Genre' data={movie?.Genre} />
-                            <DetailColumns title='Director' data={movie?.Director} />
+                        <div className="image">
+                            <img src={movie?.Poster} alt={movie?.Title} />
                         </div>
                     </div>
-                    <div className="image">
-                        <img src={movie?.Poster} alt={movie?.Title} />
-                    </div>
-                </div>
-            </>
-            }
+                </>
+            : null}
         </div>
     )
 }
