+++ my-code/src/components/detailColumns.tsx
@@ -1,5 +1,5 @@
 interface DetailColumnsProps {
-    title?: string;
+    title: string;
     data?: string;
 }
 
@@ -7,10 +7,10 @@ export default function DetailColumns({ title, data }: DetailColumnsProps) {
     return (
         <div className="plot">
             <h4 className="title">
-                Genre
+                {title}
             </h4>
-            {data?.split(',').map((item, i) => {
-                return <h4 key={i} className="content">{item}</h4>;
+            {data?.split(',').map((item, index) => {
+                return <h4 key={index} className="content">{item}</h4>;
             })}
         </div>
     )
