--- my-code/src/utils/checkResultAsError.ts
@@ -1,5 +1,0 @@
-import { ErrorResponse } from "../interfaces/resultAsError";
-
-export const resultIsError = (data: any): data is ErrorResponse => {
-    return 'Error' in data;
-}
